+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - New notification system implemented
 - Update to comply with new PV Online api 
 - Show ingredients, synonyms and compositions count when you importing from PV Online
+- Improve PV Online import and upload functions
+- Improve formula creation functions
 - TECH Preview: PV Online share With a friend
 
 ### Version 4.7